LayoutTests/http/tests/websocket/tests/hybi/reserved-opcodes-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=3' failed: Unrecognized frame opcode: 3
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=4' failed: Unrecognized frame opcode: 4
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=5' failed: Unrecognized frame opcode: 5
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=6' failed: Unrecognized frame opcode: 6
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=7' failed: Unrecognized frame opcode: 7
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=11' failed: Unrecognized frame opcode: 11
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=12' failed: Unrecognized frame opcode: 12
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=13' failed: Unrecognized frame opcode: 13
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=14' failed: Unrecognized frame opcode: 14
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/reserved-opcodes?opcode=15' failed: Unrecognized frame opcode: 15
Tests whether WebSocket ignores frames with reserved opcode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Opcode 3: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 4: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 5: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 6: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 7: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 11: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 12: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 13: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 14: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 15: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
